+++ R/Shiny.R
@@ -93,6 +93,8 @@ launchDiagnosticsExplorer <- function(dataFolder = "data",
   ensure_installed("stringr")
   ensure_installed("SqlRender")
   ensure_installed("tidyr")
+  ensure_installed("CirceR")
+  ensure_installed("rmarkdown")
   
   if (launch.browser) {
     options(shiny.launch.browser = TRUE)
@@ -301,7 +303,13 @@ ensure_installed <- function(pkg) {
     if (interactive()) {
       message(msg, "\nWould you like to install it?")
       if (menu(c("Yes", "No")) == 1) {
-        install.packages(pkg)
+        if (pkg == 'CirceR') {
+          ensure_installed("remotes")
+          message(msg, "\nInstalling from Github using remotes")
+          remotes::install_github("OHDSI/CirceR")
+        } else {
+          install.packages(pkg)
+        }
       } else {
         stop(msg, call. = FALSE)
       }