+++ extras/ExampleOfHowToCreateStudyPackageWithHydra.R
@@ -1,4 +1,6 @@
-#remotes::install_github("OHDSI/Hydra")
+# please ensure you have the latest version of Hydra. As of 08/13/2021 - CohortDiagnostics support for Hydra is still in develop branch.
+# please check hydra release notes and update hydra
+remotes::install_github("OHDSI/Hydra", ref = "develop")
 outputFolder <- "d:/temp/output"  # location where you study package will be created
 
 
@@ -18,7 +20,9 @@ description <- "Cohort diagnostics on Thrombosis With Thrombocytopenia Syndrome 
 
 library(magrittr)
 # Set up
-baseUrl <- Sys.getenv("baseUrlUnsecure")
+baseUrl <- Sys.getenv("baseUrl")
+# if you have security enabled, please authorize the use - example below
+# ROhdsiWebApi::authorizeWebApi(baseUrl, 'windows')
 cohortIds <- c(22040,
                22042,
                22041,
@@ -66,37 +70,39 @@ specifications <- list(id = 1,
                        version = version,
                        name = name,
                        packageName = packageName,
-                       skeletonVersin = skeletonVersion,
+                       skeletonVersion = skeletonVersion,
                        createdBy = createdBy,
                        createdDate = createdDate,
                        modifiedBy = modifiedBy,
                        modifiedDate = modifiedDate,
-                       skeletontype = skeletonType,
+                       skeletonType = skeletonType,
                        organizationName = organizationName,
                        description = description,
                        cohortDefinitions = cohortDefinitionsArray)
 
 jsonFileName <- paste0(file.path(tempFolder, "CohortDiagnosticsSpecs.json"))
-write(x = specifications %>% RJSONIO::toJSON(pretty = TRUE), file = jsonFileName)
+write(x = specifications %>% RJSONIO::toJSON(pretty = TRUE, digits = 23), file = jsonFileName)
 
 
 ##############################################################
 ##############################################################
 #######       Get skeleton from github            ############
+#######       Uncomment if you want to use latest ############
+#######       skeleton only - for advanced user   ############
 ##############################################################
 ##############################################################
 ##############################################################
 #### get the skeleton from github
-download.file(url = "https://github.com/OHDSI/SkeletonCohortDiagnosticsStudy/archive/refs/heads/main.zip",
-                         destfile = file.path(tempFolder, 'skeleton.zip'))
-unzip(zipfile =  file.path(tempFolder, 'skeleton.zip'), 
-      overwrite = TRUE,
-      exdir = file.path(tempFolder, "skeleton")
-        )
-fileList <- list.files(path = file.path(tempFolder, "skeleton"), full.names = TRUE, recursive = TRUE, all.files = TRUE)
-DatabaseConnector::createZipFile(zipFile = file.path(tempFolder, 'skeleton.zip'), 
-                                 files = fileList, 
-                                 rootFolder = list.dirs(file.path(tempFolder, 'skeleton'), recursive = FALSE))
+# download.file(url = "https://github.com/OHDSI/SkeletonCohortDiagnosticsStudy/archive/refs/heads/main.zip",
+#                          destfile = file.path(tempFolder, 'skeleton.zip'))
+# unzip(zipfile =  file.path(tempFolder, 'skeleton.zip'), 
+#       overwrite = TRUE,
+#       exdir = file.path(tempFolder, "skeleton")
+#         )
+# fileList <- list.files(path = file.path(tempFolder, "skeleton"), full.names = TRUE, recursive = TRUE, all.files = TRUE)
+# DatabaseConnector::createZipFile(zipFile = file.path(tempFolder, 'skeleton.zip'), 
+#                                  files = fileList, 
+#                                  rootFolder = list.dirs(file.path(tempFolder, 'skeleton'), recursive = FALSE))
 
 ##############################################################
 ##############################################################
@@ -111,9 +117,14 @@ hydraSpecificationFromFile <- Hydra::loadSpecifications(fileName = jsonFileName)
 unlink(x = outputFolder, recursive = TRUE)
 dir.create(path = outputFolder, showWarnings = FALSE, recursive = TRUE)
 Hydra::hydrate(specifications = hydraSpecificationFromFile,
-               outputFolder = outputFolder, 
-               skeletonFileName = file.path(tempFolder, 'skeleton.zip')
+               outputFolder = outputFolder
 )
+
+# for advanced user using skeletons outside of Hydra
+# Hydra::hydrate(specifications = hydraSpecificationFromFile,
+#                outputFolder = outputFolder, 
+#                skeletonFileName = file.path(tempFolder, 'skeleton.zip')
+# )
 
 
 unlink(x = tempFolder, recursive = TRUE, force = TRUE)