+++ NEWS.md
@@ -1,3 +1,31 @@
+CohortDiagnostics 2.1.3
+=======================
+
+Changes:
+1. On starting diagnostics explorer using launchDiagnosticsExplorer - checks were added for remotes and CirceR https://github.com/OHDSI/CohortDiagnostics/issues/595
+
+Bug Fixes:
+1. Diagnostics explorer - characterization plot would show a warning message when one of the cohorts selected has no data. Added check for this issue.
+
+CohortDiagnostics 2.1.2
+=======================
+
+Bug fixes:
+1. DiagnosticsExplorer fixes a bug of app failure when runIndexEventBreakdown, runOrphanConcepts, includedSourceConcept is set to FALSE
+
+
+CohortDiagnostics 2.1.1
+=======================
+
+Bug fixes:
+
+1. DiagnosticsExplorer shiny app bug fixes: inclusion rule display. 
+2. Handle situation where cdm_source may have more than one row record. warning + use max value for vocabularyVersion.
+3. Switch off runTimeSeries for BigQuery because of SQL translation issue, to be addressed in version 2.3. https://github.com/OHDSI/CohortDiagnostics/issues/503 Also runTimeSeries is set to FALSE by default.
+3. Fixed a bug in DDL - temporal_covariate_value_dist https://github.com/OHDSI/CohortDiagnostics/issues/490
+See issue https://github.com/OHDSI/CohortDiagnostics/issues/497
+
+
 CohortDiagnostics 2.1.0
 =======================